refactor models location and naming
