+++ backend/backend.api/Controllers/InfoOrganizationController.cs
@@ -1,5 +1,5 @@
 ﻿using AutoMapper;
-using backend.api.Models;
+using backend.api.Models.RequestModels;
 using backend.servicios.DTOs;
 using backend.servicios.Interfaces;
 using Microsoft.AspNetCore.Mvc;
