+++ backend/backend.api/Controllers/MapsController.cs
@@ -1,5 +1,5 @@
 ﻿using AutoMapper;
-using backend.api.Models;
+using backend.api.Models.ResponseModels;
 using backend.servicios.Interfaces;
 using Microsoft.AspNetCore.Mvc;
