+++ backend/backend.api/Controllers/NeedController.cs
@@ -1,5 +1,5 @@
 ﻿using AutoMapper;
-using backend.api.Models;
+using backend.api.Models.ResponseModels;
 using backend.servicios.Interfaces;
 using Microsoft.AspNetCore.Mvc;
