+++ backend/backend.api/Mappers/DonationProfile.cs
@@ -1,5 +1,5 @@
 ﻿using AutoMapper;
-using backend.api.Models;
+using backend.api.Models.RequestModels;
 using backend.data.Models;
 using backend.servicios.DTOs;
