+++ backend/backend.api/Mappers/IdeaProfile.cs
@@ -1,6 +1,6 @@
 ﻿using AutoMapper;
-using backend.api.Models;
 using backend.api.Models.RequestModels;
+using backend.api.Models.ResponseModels;
 using backend.data.Models;
 using backend.servicios.DTOs;
