+++ backend/backend.api/Mappers/NeedsProfile.cs
@@ -1,5 +1,5 @@
 ﻿using AutoMapper;
-using backend.api.Models;
+using backend.api.Models.ResponseModels;
 using backend.data.Models;
 using backend.servicios.DTOs;
