+++ backend/backend.api/Mappers/OrganizationProfile.cs
@@ -1,5 +1,7 @@
 ﻿using AutoMapper;
 using backend.api.Models;
+using backend.api.Models.RequestModels;
+using backend.api.Models.ResponseModels;
 using backend.data.Models;
 using backend.servicios.DTOs;
 
@@ -40,9 +42,9 @@ namespace backend.api.Mappers
 
             CreateMap<HeadquartersDto, HeadquartersResponseModel>();
 
-            CreateMap<Organizacion, HeadquartersNearby>();
+            CreateMap<Organizacion, HeadquartersNearbyDto>();
 
-            CreateMap<DataRequestModel, HeadquartersNearby>();
+            CreateMap<DataRequestModel, HeadquartersNearbyDto>();
         }
     }
 }
