+++ backend/backend.api/Models/RequestModels/ImageClassificationRequestModel.cs
@@ -1,8 +1,8 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace backend.api.Models
+namespace backend.api.Models.RequestModels
 {
-    public class BaseRequest
+    public class ImageClassificationRequestModel
     {
         [Required]
         public IFormFile Image { get; set; }
