+++ backend/backend.api/Models/RequestModels/InfoOrganizationRequest.cs
@@ -1,4 +1,4 @@
-﻿namespace backend.api.Models
+﻿namespace backend.api.Models.RequestModels
 {
     public class InfoOrganizationRequest
     {
