+++ backend/backend.api/Models/RequestModels/UpdateDonationsStateRequest.cs
@@ -1,4 +1,4 @@
-﻿namespace backend.api.Models
+﻿namespace backend.api.Models.RequestModels
 {
     public class UpdateDonationsStateRequest
     {
