+++ backend/backend.api/Models/ResponseModels/HeadquartersResponseModel.cs
@@ -1,6 +1,6 @@
 ﻿using backend.data.Models;
 
-namespace backend.api.Models
+namespace backend.api.Models.ResponseModels
 {
     public class HeadquartersResponseModel
     {
