+++ backend/tests/backend.api.test/IdeaControllerTests.cs
@@ -1,7 +1,8 @@
 ﻿using AutoMapper;
 using backend.api.Controllers;
 using backend.api.Mappers;
-using backend.api.Models;
+using backend.api.Models.RequestModels;
+using backend.api.Models.ResponseModels;
 using backend.servicios.Interfaces;
 using backend.servicios.Models;
 using Microsoft.AspNetCore.Mvc;
