+++ backend/tests/backend.api.test/MapperTests/IdeaProfileTest.cs
@@ -1,6 +1,6 @@
 ﻿using AutoMapper;
 using backend.api.Mappers;
-using backend.api.Models;
+using backend.api.Models.ResponseModels;
 using backend.data.Models;
 using backend.servicios.DTOs;
