+++ backend/tests/backend.api.test/MapperTests/UserProfileTest.cs
@@ -1,6 +1,7 @@
 ﻿using AutoMapper;
 using backend.api.Mappers;
-using backend.api.Models;
+using backend.api.Models.RequestModels;
+using backend.api.Models.ResponseModels;
 using backend.data.Models;
 using backend.servicios.DTOs;
