+++ backend/tests/backend.api.test/UsuarioRequestModelValidatorTests.cs
@@ -1,4 +1,4 @@
-﻿using backend.api.Models;
+﻿using backend.api.Models.RequestModels;
 using backend.api.Validadores;
 
 namespace backend.api.test
